Added major music commands

+++ src/commands/music/play.js
@@ -19,38 +19,39 @@ class MusicCommand extends Command {
 
     async exec(message, { search }) {
         await message.delete();
-        let guildQueue = this.client.musicPlayer.getQueue(message.guild.id);
-        let queue;
-        if (guildQueue){
-            queue = guildQueue;
-        }else{
-            queue = this.client.musicPlayer.createQueue(message.guild.id);
-        }
-        let musicEmbed = this.client.functions.embed();
-        await queue.join(message.member.voice.channel);
-        const loading = await message.channel.send("Recherche de la musique ...");
-        await queue.play(search).then(async data => {
-            setTimeout(() => {
-                if (data){
-                    loading.delete();
-                    musicEmbed.setTitle(data.name)
-                        .setURL(data.url)
-                        .setThumbnail(data.thumbnail)
-                        .setColor(this.client.colors.color.darkpurple)
-                        .setDescription(`Musique demandée par <@${message.author.id}>`)
-                        .addField("Durée", data.duration, true)
-                        .addField("Statut", data.isLive ? "En Live" : "Vidéo", true)
-                    message.channel.send({ embeds : [musicEmbed]})
-                    const ProgressBar = queue.createProgressBar();
-                    console.log(ProgressBar.prettier);
-                }else{
-                    loading.delete();
-                    message.channel.send("Aucune musique n'a été trouvé!")
+        const channel = await this.client.functions.checkMusicChannelExistence(message, message.guild, this.client.guildSettings, this.client);
+        if (channel.allowed){
+            let guildQueue = this.client.musicPlayer.getQueue(message.guild.id);
+            let queue;
+            if (guildQueue){
+                queue = guildQueue;
+            }else{
+                queue = this.client.musicPlayer.createQueue(message.guild.id);
+                await queue.setData({voiceChannelID: message.member.voice.channel.id});
+            }
+            if (await this.client.functions.checkUserInVoiceChannel(message)){
+                if (await this.client.functions.checkUserInSameVoiceChannelAsBot(message, queue)){
+                    let musicEmbed = this.client.functions.embed();
+                    await queue.join(message.member.voice.channel);
+                    const loading = await channel.channelObject.send("Recherche de la musique ...");
+                    await queue.play(search).then(async data => {
+                        setTimeout(() => {
+                            if (data){
+                                loading.delete();
+                                musicEmbed = this.client.functions.musicEmbed(data.name, data.url, data.duration, data.isLive, data.thumbnail, message.author, this.client.colors.color.darkpurple, queue.songs.length)
+                                channel.channelObject.send({ embeds : [musicEmbed]})
+                            }else{
+                                loading.delete();
+                                channel.channelObject.send("Aucune musique n'a été trouvé!")
+                            }
+                        },2000)
+                    }).catch(error => {
+                        channel.channelObject.send(error.message);
+                    })
                 }
-            },2000)
-        }).catch(error => {
-            message.channel.send(error.message);
-        })
+            }
+        }
+
     }
 }
 module.exports = MusicCommand;
